gem: create user-space mappable BOs as uncached
User-space use mappable BOs notably for fences, and expects that a
value update by the GPU will be immediatly visible through the
user-space mapping.

ARM has a property that may prevent this from happening though: memory
can be mapped multiple times only if the different mappings share the
same caching properties. However all the lowmem memory is already
identity-mapped into the kernel with cache enabled, so when user-space
requests an uncached mapping, we actually get an "undefined caching
policy" one and this has strange side-effects described on Freedesktop
bug 86690.

To prevent this from happening, create objects that are requested to be
mappable by user-space with TTM_PL_FLAG_UNCACHED. This will request
Nouveau to allocate pages using the DMA API, which will fix the identify
mapping and allow us to safely map the objects to user-space uncached.

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>